Handle calls to epoll_* and operate on any internally used sockets
This is the bulk of the implementation. On a call to epoll_create,
we initialize evsocks, if it has not be done, or we add a new struct to
the linked-list and set its values, we then return the new evfd.

On a call to epoll_ctl, if the op is ADD, then we set event->data->fd
to be fd if it's not already the socket's file descriptor, and we add
it to the event-based file descriptors's linked-list of socket file
descriptors. If it's MOD or DEL then just do it, and reset fd if
necessary.

On a call to epoll_wait, retrieve the new events and check if any of them are
for us. If so, take action on them. Once complete, remove all of the events
that were ours from the array and return only that number of events. Reset
event->data->fd to the value the application expects if we modified it
earlier.

On close of any of the fds we are tracking free the entries in the
linked-list.