Add support for AF_UNIX SOCKS connections
Torsocks now recognizes unix:/path/to/sock file paths
as a Unix domain socket pathname. AF_UNIX application connection
handling is not changed.

Now when an application requests a new socket of a specified type
it returns a valid fd for that address family and uses it as a unique
key in the connection hash table. When this fd is used in a subsequent
libc call the connection is retrieved and the appropriate socket is used
when connecting to Tor's socks port.