For internal tor connections alloc conn on the heap
We previously allocated it on the stack but that now causes a
problem when we close the connection fd and then try to free
the conn struct (oops). If we put this on the heap, like we do
when the application asks for the connection, then we can handle
this exactly the same and let the existing logic free the struct
without any problem.